streaming/producers/requirements.txt
kafka-python==2.0.2
faker==19.12.0
pytz==2023.3
python-dateutil==2.8.2 